move out spell-storage
