+++ Cargo.toml
@@ -37,6 +37,7 @@ members = [
     "particle-modules",
     "connection-pool",
     "script-storage",
+    "spell-storage",
     "particle-execution",
 ]
 exclude = [
@@ -60,3 +61,4 @@ fluence-keypair = "0.8.1"
 parking_lot = "0.12.1"
 async-std = { version = "1.12.0", features = ["unstable"] }
 uuid = { version = "1.2.1", features = ["v4"] }
+derivative = "2.2.0"
