+++ spell-storage/src/lib.rs
@@ -1,0 +1,3 @@
+mod storage;
+
+pub use crate::storage::SpellStorage;
